+++ app/models/challenge.rb
@@ -1,6 +1,6 @@
 class Challenge < ActiveRecord::Base
-    has_many :user_challenges
-    has_many :user_challenge_completeds
+    has_many :user_challenges, :dependent => :destroy
+    has_many :user_challenge_completeds, :dependent => :destroy
     has_many :current_users, through: :user_challenges, class_name: "User", foreign_key: "user_id", source: :user
     has_many :completed_users, through: :user_challenge_completeds, class_name: "User", foreign_key: "user_id", source: :user
     validates :name, presence: true
@@ -49,10 +49,16 @@ class Challenge < ActiveRecord::Base
             challengeObj = Challenge.find(id)
             if challengeObj
                 if delete
-                    challengeObj.destroy
-                    status = true
-                    message = "Challenge deleted"
-                    elements = nil
+                    if challengeObj.user_challenge_completeds.length == 0
+                        challengeObj.destroy
+                        status = true
+                        message = "Challenge deleted"
+                        elements = nil
+                    else
+                        status = false
+                        message = "#{challengeObj.name} cannot be deleted: already completed by a user."
+                        elements = nil
+                    end
                 else
                     # evaluate the criteria to make sure it's actually good
                     evaluation_result = criteria == challengeObj.criteria ? true : User.find_users_matching_criteria(criteria)[:status]
@@ -67,7 +73,7 @@ class Challenge < ActiveRecord::Base
                             savedObj = challengeObj.save
                         rescue => error
                             status = false
-                            message = "Challenge could not be updated: #{error.message}"
+                            message = "#{challengeObj.name} could not be updated: #{error.message}"
                             elements = nil
                         else
                             if savedObj
@@ -76,7 +82,7 @@ class Challenge < ActiveRecord::Base
                                 elements = nil
                             else
                                 status = false
-                                message = "Challenge could not be updated: #{challengeObj.errors.full_messages.join(', ')}"
+                                message = "#{challengeObj.name} could not be updated: #{challengeObj.errors.full_messages.join(', ')}"
                                 elements = challengeObj.errors.messages.keys
                             end
                         end
